+++ src/commands/create.ts
@@ -5,7 +5,7 @@ import path from "node:path";
 
 import { command } from "../common/command";
 import { createCopier } from "../common/file-copier";
-import { invalidCharacters, isExecutable, isSafeBaseName, relative } from "../common/path";
+import { invalidCharacters, isExecutable, isSafeBaseName, relative } from "../system/path";
 import { run } from "../common/runner";
 import { StdOut } from "../common/stdout";
 import { getConfig } from "../config";